pyproject.toml
[project]
name = "webscale-multimodal-datapipeline"
version = "0.1.0"
description = "High-performance, distributed web-scale multimodal data processing pipeline with Ray, featuring Rust-accelerated and GPU-optimized operators"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Duo An", email = "duoan@example.com" }]
keywords = [
    "data-pipeline",
    "multimodal",
    "ray",
    "distributed",
    "image-processing",
    "clip",
    "siglip",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "datasets>=4.5.0",
    "ffmpeg-python>=0.2.0",
    "imagehash>=4.3.2",
    "numpy>=1.24.0",
    "pillow>=12.1.0",
    "pyarrow>=14.0.0",
    "pyiceberg>=0.7.0",
    "pyyaml>=6.0.0",
    "ray[data,default]>=2.53.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "open-clip-torch>=3.2.0",
    "huggingface-hub>=0.20.0",
    "safetensors>=0.4.0",
    "transformers>=4.57.6",
]

[project.optional-dependencies]
enhanced = [
    "langdetect>=1.0.9",    # For better language detection
    "pytesseract>=0.3.10",  # For OCR text detection
    "opencv-python>=4.8.0", # For advanced image processing
]
ml = [
    "torch>=2.0.0",            # PyTorch for deep learning models
    "torchvision>=0.15.0",     # Torchvision for image processing
    "open-clip-torch>=2.20.0", # OpenCLIP for image embeddings
]

[project.urls]
Homepage = "https://github.com/duoan/webscale-multimodal-datapipeline"
Repository = "https://github.com/duoan/webscale-multimodal-datapipeline"
Documentation = "https://github.com/duoan/webscale-multimodal-datapipeline#readme"

[project.scripts]
wmd = "webscale_multimodal_datapipeline.cli:main"

[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "."
module-name = "webscale_multimodal_datapipeline.rust_accelerated_ops"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["webscale_multimodal_datapipeline"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
